+++ examples/napi/__tests__/values.spec.ts
@@ -734,7 +734,7 @@ test('Async error with stack trace', async (t) => {
   t.not(err?.stack, undefined)
   t.deepEqual(err!.message, 'Async Error')
   if (!process.env.WASI_TEST) {
-    t.regex(err!.stack!, /.+at .+values\.spec\.ts:\d+:\d+.+/gm)
+    t.regex(err!.stack!, /.+at .+values\.spec\.(ts|js):\d+:\d+.+/gm)
   }
 })
 
@@ -796,7 +796,7 @@ test('aliased rust struct and enum', (t) => {
 })
 
 test('serde-json', (t) => {
-  if (process.env.WASI_TEST) {
+  if (process.env.WASI_TEST || process.platform === 'freebsd') {
     t.pass()
     return
   }
@@ -1234,9 +1234,10 @@ Napi4Test('throw error from ThreadsafeFunction', async (t) => {
 
 Napi4Test('ThreadsafeFunction closure capture data', (t) => {
   return new Promise((resolve) => {
-    threadsafeFunctionClosureCapture(() => {
+    const defaultValue = new Animal(Kind.Dog, '旺财')
+    threadsafeFunctionClosureCapture(defaultValue, (value) => {
       resolve()
-      t.pass()
+      t.is(value, defaultValue)
     })
   })
 })
@@ -1546,6 +1547,7 @@ test('create readable stream from channel', async (t) => {
   }
   t.is(Buffer.concat(chunks).toString('utf-8'), 'hello'.repeat(100))
   const { ReadableStream } = await import('web-streams-polyfill')
+  // @ts-expect-error polyfill ReadableStream is not the same as the one in Node.js
   const streamFromClass = await createReadableStreamFromClass(ReadableStream)
   const chunksFromClass = []
   for await (const chunk of streamFromClass) {
@@ -1570,7 +1572,7 @@ test('spawnThreadInThread should be fine', async (t) => {
 })
 
 test('should generate correct type def file', async (t) => {
-  if (process.env.WASI_TEST) {
+  if (process.env.WASI_TEST || process.platform === 'freebsd') {
     t.pass()
   } else {
     t.snapshot(await nodeReadFile(join(__dirname, '..', 'index.d.cts'), 'utf8'))